Display sync method
